+++ src/main/java/it/challenging/torchy/entity/NeedModificato.java
@@ -27,7 +27,7 @@ public class NeedModificato implements Serializable {
 
     private String descrizione;
 
-    private Integer anniEsperienza;
+    private Double anniEsperienza;
 
     private Integer numeroRisorse;
